compile with all flags

+++ sources/get_env_var_value.c
@@ -6,7 +6,7 @@
 /*   By: ale-baux <ale-baux@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/06 13:37:57 by ale-baux          #+#    #+#             */
-/*   Updated: 2020/03/06 15:52:02 by ale-baux         ###   ########.fr       */
+/*   Updated: 2020/03/12 18:18:06 by ale-baux         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ int is_variable(char *var_name, char *var_line)
 	int len;
 
 	len = ft_strlen(var_name);
-	if (len > ft_strlen(var_line))
+	if (len > (int)ft_strlen(var_line))
 		return (0);
 	len--;
 	while (len >= 0)
